+++ src/Field/Checkbox.php
@@ -69,10 +69,11 @@ class Checkbox extends \Geniem\ACF\Field {
      * Export field in ACF's native format.
      *
      * @param boolean $register Whether the field is to be registered.
+     * @param mixed   $parent   Possible parent object.
      *
      * @return array
      */
-    public function export( bool $register = false ) : ?array {
+    public function export( bool $register = false, $parent = null ) : ?array {
         if ( $register ) {
             if ( $this->disable_all ) {
                 $this->disabled = array_keys( $this->choices );
@@ -82,7 +83,7 @@ class Checkbox extends \Geniem\ACF\Field {
         }
 
         // Call the original export method
-        $obj = parent::export( $register );
+        $obj = parent::export( $register, $parent );
 
         return $obj;
     }